DEV-39. BlogPagination => Pagination (Now it's common component). Now Pagination component could be controlled outside (added props)
