--- src/components/BlogPagination/index.tsx
@@ -1,3 +1,0 @@
-import BlogPagination from './BlogPagination'
-
-export default BlogPagination;
