+++ src/components/Pagination/Pagination.scss
@@ -1,3 +1,3 @@
-.blog__pagination{
+.pagination{
    text-align: center;
 }
