+++ src/components/Pagination/Pagination.test.tsx
@@ -1,11 +1,13 @@
-import BlogPagination from './BlogPagination';
+import Pagination from './Pagination';
 import renderer from 'react-test-renderer';
 
 describe('BlogPagination', () => {
    it('Should successfully render', () => {
       const wrapper = renderer
          .create(
-            <BlogPagination />
+            <Pagination
+               defaultCurrentPage={3}
+               totalItems={500} />
          )
          .toJSON();
       expect(wrapper).toMatchSnapshot();
